Move fake minitest controller test case to file
