+++ spec/support/fake_minitest_controller_test_case.rb
@@ -1,0 +1,26 @@
+# This class replicates the behaviour of a `ActionController::TestCase` to allow
+# us to test the minitest helpers, in RSpec.
+class FakeMinitestControllerTestCase
+  def initialize
+    @request = FakeRequestResponseObject.new
+  end
+
+  def assert_equal(*)
+  end
+
+  def assert_select(*)
+  end
+
+  def assert_match(*)
+  end
+
+  def response
+    FakeRequestResponseObject.new
+  end
+
+  class FakeRequestResponseObject
+    def headers
+      {}
+    end
+  end
+end
